requirements.txt
flask
lask_cors
requests
shutil
googlesearch